refactor(console): multi-card selector
