+++ packages/console/src/pages/CloudPreview/components/CardSelector/index.tsx
@@ -1,27 +1,3 @@
-import type { AdminConsoleKey } from '@logto/phrases';
-import type { ReactNode } from 'react';
-
-import RadioGroup, { Radio } from '@/components/RadioGroup';
-
-export type Option = {
-  icon?: ReactNode;
-  title: AdminConsoleKey;
-  value: string;
-};
-
-type Props = {
-  name: string;
-  value: string;
-  options: Option[];
-  onChange: (value: string) => void;
-};
-
-const CardSelector = ({ name, value, options, onChange }: Props) => (
-  <RadioGroup type="compact" value={value} name={name} onChange={onChange}>
-    {options.map(({ value: optionValue, title, icon }) => (
-      <Radio key={optionValue} icon={icon} title={title} value={optionValue} />
-    ))}
-  </RadioGroup>
-);
-
-export default CardSelector;
+export type { Option } from './types';
+export { default as CardSelector } from './CardSelector';
+export { default as MultiCardSelector } from './MultiCardSelector';
