+++ packages/console/src/pages/CloudPreview/components/CardSelector/types.ts
@@ -1,0 +1,8 @@
+import type { AdminConsoleKey } from '@logto/phrases';
+import type { ReactNode } from 'react';
+
+export type Option = {
+  icon?: ReactNode;
+  title: AdminConsoleKey;
+  value: string;
+};
